requirements.txt
fastapi==0.115.4
uvicorn==0.32.1
python-multipart==0.0.12
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==1.56.2
requests==2.32.3
httpx==0.28.1
pydantic==2.10.3
